+++ EssentialDevApp/EssentialDevAppTests/Helpers/InMemoryFeedStore.swift
@@ -1,0 +1,55 @@
+//
+//  InMemoryFeedStore.swift
+//  EssentialDevAppTests
+//
+//  Created by Min-Yang Huang on 5/6/25.
+//
+
+import Foundation
+import EssentialDev
+
+class InMemoryFeedStore: FeedStore, FeedImageDataStore {
+    private(set) var feedCache: CachedFeed?
+    private var feedImageDataCache: [URL: Data] = [:]
+    
+    private init(feedCache: CachedFeed? = nil) {
+        self.feedCache = feedCache
+    }
+    
+    func deleteCachedFeed(completion: @escaping FeedStore.DeletionCompletion) {
+        feedCache = nil
+        completion(.success(()))
+    }
+    
+    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping FeedStore.InsertionCompletion) {
+        feedCache = CachedFeed(feed: feed, timestamp: timestamp)
+        completion(.success(()))
+    }
+    
+    func retrieve(completion: @escaping FeedStore.RetrievalCompletion) {
+        completion(.success(feedCache))
+    }
+    
+    func insert(_ data: Data, for url: URL, completion: @escaping (FeedImageDataStore.InsertionResult) -> Void) {
+        feedImageDataCache[url] = data
+        completion(.success(()))
+    }
+    
+    func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
+        completion(.success(feedImageDataCache[url]))
+    }
+}
+
+extension InMemoryFeedStore {
+    static var empty: InMemoryFeedStore {
+        InMemoryFeedStore()
+    }
+    
+    static var withExpiredFeedCache: InMemoryFeedStore {
+        InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date.distantPast))
+    }
+    
+    static var withNonExpiredFeedCache: InMemoryFeedStore {
+        InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date()))
+    }
+}
